move home page components to components/home folder
