+++ apps/frontend/src/components/home/calendar-section.tsx
