+++ apps/frontend/src/components/home/hero-section.tsx
