+++ apps/frontend/src/components/home/programs-section.tsx
