+++ apps/frontend/src/components/home/video-section.tsx
